Move custom_codecs.py and custom_json.py to the json serialization subpackage
