+++ gufe/serialization/json/custom_codecs.py
@@ -12,7 +12,7 @@ from openff.units import DEFAULT_UNIT_REGISTRY
 
 import gufe
 from gufe.compression import zst_compress, zst_decompress
-from gufe.custom_json import JSONCodec
+from gufe.serialization.json.custom_json import JSONCodec
 from gufe.settings.models import SettingsBaseModel
